new-state: document new states
